Move all tables to page files.
