+++ static/js/init.js
@@ -23,7 +23,6 @@ var typeahead = require('./modules/typeahead.js');
 var charts = require('./modules/charts.js');
 var glossary = require('./modules/glossary.js');
 var Search = require('./modules/search');
-var tables = require('./modules/tables');
 var maps = require('./modules/maps');
 var toggle = require('./modules/toggle');
 
@@ -175,6 +174,5 @@ $(document).ready(function() {
     // @endif
 
     filters.init();
-    tables.init();
     toggle.init();
 });
